MaggyDriver now has an auto-increment PK column, id
